+++ src/main/java/dev/mapselect/registry/MapSelectBlocks.java
@@ -40,6 +40,13 @@ public final class MapSelectBlocks {
 		.nonOpaque()
 		.pistonBehavior(PistonBehavior.DESTROY));
 
+	public static final Block IWY_PLUSH = new GreyiferPlushBlock(AbstractBlock.Settings.create()
+		.mapColor(MapColor.WHITE)
+		.strength(0.5f)
+		.sounds(BlockSoundGroup.WOOL)
+		.nonOpaque()
+		.pistonBehavior(PistonBehavior.DESTROY));
+
 	public static final Block FAKE_SUSPICIOUS_SAND = new Block(AbstractBlock.Settings.create()
 		.mapColor(MapColor.PALE_YELLOW)
 		.strength(0.25f)
@@ -62,6 +69,7 @@ public final class MapSelectBlocks {
 	public static final BlockItem SAND_LAYER_ITEM = new BlockItem(SAND_LAYER, new Item.Settings());
 	public static final BlockItem RED_SAND_LAYER_ITEM = new BlockItem(RED_SAND_LAYER, new Item.Settings());
 	public static final BlockItem GREYIFER_PLUSH_ITEM = new BlockItem(GREYIFER_PLUSH, new Item.Settings());
+	public static final BlockItem IWY_PLUSH_ITEM = new BlockItem(IWY_PLUSH, new Item.Settings());
 	public static final BlockItem FAKE_SUSPICIOUS_SAND_ITEM = new BlockItem(FAKE_SUSPICIOUS_SAND, new Item.Settings());
 	public static final BlockItem FAKE_SUSPICIOUS_GRAVEL_ITEM = new BlockItem(FAKE_SUSPICIOUS_GRAVEL, new Item.Settings());
 	public static final BlockItem PEBBLE_BLOCK_ITEM = new BlockItem(PEBBLE_BLOCK, new Item.Settings());
@@ -70,18 +78,21 @@ public final class MapSelectBlocks {
 		Identifier sandId = Identifier.of(MapSelect.MOD_ID, "sand_layer");
 		Identifier redSandId = Identifier.of(MapSelect.MOD_ID, "red_sand_layer");
 		Identifier greyiferPlushId = Identifier.of(MapSelect.MOD_ID, "greyifer_plush");
+		Identifier iwyPlushId = Identifier.of(MapSelect.MOD_ID, "iwy_plush");
 		Identifier fakeSuspiciousSandId = Identifier.of(MapSelect.MOD_ID, "fake_suspicious_sand");
 		Identifier fakeSuspiciousGravelId = Identifier.of(MapSelect.MOD_ID, "fake_suspicious_gravel");
 		Identifier pebbleBlockId = Identifier.of(MapSelect.MOD_ID, "pebble_block");
 		Registry.register(Registries.BLOCK, sandId, SAND_LAYER);
 		Registry.register(Registries.BLOCK, redSandId, RED_SAND_LAYER);
 		Registry.register(Registries.BLOCK, greyiferPlushId, GREYIFER_PLUSH);
+		Registry.register(Registries.BLOCK, iwyPlushId, IWY_PLUSH);
 		Registry.register(Registries.BLOCK, fakeSuspiciousSandId, FAKE_SUSPICIOUS_SAND);
 		Registry.register(Registries.BLOCK, fakeSuspiciousGravelId, FAKE_SUSPICIOUS_GRAVEL);
 		Registry.register(Registries.BLOCK, pebbleBlockId, PEBBLE_BLOCK);
 		Registry.register(Registries.ITEM, sandId, SAND_LAYER_ITEM);
 		Registry.register(Registries.ITEM, redSandId, RED_SAND_LAYER_ITEM);
 		Registry.register(Registries.ITEM, greyiferPlushId, GREYIFER_PLUSH_ITEM);
+		Registry.register(Registries.ITEM, iwyPlushId, IWY_PLUSH_ITEM);
 		Registry.register(Registries.ITEM, fakeSuspiciousSandId, FAKE_SUSPICIOUS_SAND_ITEM);
 		Registry.register(Registries.ITEM, fakeSuspiciousGravelId, FAKE_SUSPICIOUS_GRAVEL_ITEM);
 		Registry.register(Registries.ITEM, pebbleBlockId, PEBBLE_BLOCK_ITEM);
@@ -93,15 +104,20 @@ public final class MapSelectBlocks {
 			entries.add(FAKE_SUSPICIOUS_GRAVEL_ITEM);
 			entries.add(PEBBLE_BLOCK_ITEM);
 		});
-		ItemGroupEvents.modifyEntriesEvent(ItemGroups.FUNCTIONAL).register(entries -> entries.add(GREYIFER_PLUSH_ITEM));
+		ItemGroupEvents.modifyEntriesEvent(ItemGroups.FUNCTIONAL).register(entries -> {
+			entries.add(GREYIFER_PLUSH_ITEM);
+			entries.add(IWY_PLUSH_ITEM);
+		});
 		ItemGroupEvents.modifyEntriesEvent(ItemGroups.SEARCH).register(entries -> {
 			entries.add(GREYIFER_PLUSH_ITEM);
+			entries.add(IWY_PLUSH_ITEM);
 			entries.add(FAKE_SUSPICIOUS_SAND_ITEM);
 			entries.add(FAKE_SUSPICIOUS_GRAVEL_ITEM);
 			entries.add(PEBBLE_BLOCK_ITEM);
 		});
 		ItemGroupEvents.modifyEntriesEvent(WatheItems.DECORATION_GROUP).register(entries -> {
 			entries.add(GREYIFER_PLUSH_ITEM);
+			entries.add(IWY_PLUSH_ITEM);
 			entries.add(FAKE_SUSPICIOUS_SAND_ITEM);
 			entries.add(FAKE_SUSPICIOUS_GRAVEL_ITEM);
 			entries.add(PEBBLE_BLOCK_ITEM);